requirements.txt
python-telegram-bot==21.6
aiosqlite==0.20.0
python-dotenv==1.0.1
telethon==1.42.0
pytgcalls
tgcrypto
av
